+++ glibc/string/bzero.c
@@ -3,8 +3,7 @@
 int main(void)
 {
 	char buf[1024] = { 'x' };
-	char buf2[1024] = { 'z' };
 
-	bcopy(buf, buf2, sizeof(buf));
+	bzero(buf, sizeof(buf));
 	return 0;
 }